[REF] move definition of CValue_array to a separate cpp file
